+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Semantic versioning is used to denote different versions of this project.
 ### Added
 ### Changed
 - Split the concrete cover input into separate fields for top and bottom cover values, allowing for different cover specifications for the top and bottom of the plate.
+- Split the shear reinforcement input fields into separate fields for top and bottom.
 ### Fixed
 
 ## [`v0.0.5`] - 2025-05-22